Extract Tests::Cleanup
Next part of splitting Test into separate classes.
